Move Cli extension and rpc implementation to separate modules
